Modified ISRs to accept "gpio pin #" parameter.  Optimized ISR handling.

+++ stm32/navx-mxp/navx-mxp_hal.h
@@ -155,11 +155,11 @@ int HAL_IOCX_Ext_Power_Fault(); /* Returns 0 if no ext power fault has occurred,
 void HAL_IOCX_RPI_GPIO_Driver_Enable(int enable);
 void HAL_IOCX_RPI_COMM_Driver_Enable(int enable);
 void HAL_IOCX_Init();
-void HAL_IOCX_AssertInterrupt(uint32_t int_bits_to_set);
-void HAL_IOCX_DeassertInterrupt(uint32_t int_bits_to_clear);
-void HAL_IOCX_UpdateInterruptMask(uint32_t int_new_mask);
-uint32_t HAL_IOCX_GetInterruptMask();
-uint32_t HAL_IOCX_GetInterruptStatus();
+void HAL_IOCX_AssertInterrupt(uint16_t int_bits_to_set);
+void HAL_IOCX_DeassertInterrupt(uint16_t int_bits_to_clear);
+void HAL_IOCX_UpdateInterruptMask(uint16_t int_new_mask);
+uint16_t HAL_IOCX_GetInterruptMask();
+uint16_t HAL_IOCX_GetInterruptStatus();
 #endif
 
 /* Timers (QuadEncoder/PWM) */
